More UI & Map Logic changes
1. The map drawing inside it's layout should had not a FIXED size in pixels, it was depending on the dimension of the noiseMatrix resulted from the Map Pane's width/height (in pixels) divided into the variable tileSize selected before. Considering the old resolution for map was: 1280*900, for example, 3 does not divide 1280 or 7 does not divide 900! We would like to keep the tileSizes in pixels like that.
Solution: Map Pane's (in pixels) should be choosen in such way that it will be able to be EXACTLY divided by each of the tileSizes in pixels: 3,4,5,6,7 (depending on the map type selected before). So we are doing the least common multiple of (3,4,5,6,7)=420 -> a close resolution of 1260*840(3:2 aspect ratio) for the Map Pane should work fine for the 1600x900 window size we test on. Soon we will have a responsive update for >=720p monitors.
Why? Doing so we'll have noiseMatrix's width=1260/tileSize and height=840/tileSize (both INTEGER numbers, better than doubles). Why? An unexact division with the tileSizes would result in a double width and height, that must be converted into int in order to be the number of columns/rows in the matrixes. Conversion Double to Int would result in precision loss and different sizes for the mapPane's when drawn.
Bad example, tilesize=7: 1280/7=182.8571428571429 -> converted to 182 and 900/7 -> 128.5714285714286 converted to 128
now multiplying those by the tile drawing size we have 182*7=1274 and 128*7=896 a totally different resolutions than the one we chose!

2. Few Changes to the UI: added more containers for easy alignment. Following commit will be more focused on this part.