New introUI & graphics using CSS
The introUI was totally remaked using layouts. The structure is done via Java Code, while the graphics is now styled with an external css.
This allows separation of the UI structure/graphics, but also permits flexibility for the game's graphics, as long as it fits to the structure in the java code.