Buildings System Implemented successfully.
Players are having now a winning objetcive(a set of need to complete buildings), that the Game class will decide.

For each objective building in that set, the players will try to build it by fulfilling their build requirements. The aproach is top-down as it follows: every building may have another already owned building requirements or resources cost requirements. If players do fulfill both of theese requirements, the player will build that building in his neighbourhood(if free), or if the neighbourhood is already occupied, the player will randomly teleport in another free-neighbour-spot to build the building. If the player does not fulfill the requirements, he will divide and conquer the requirements and call the build method for every building in those requirements, or if the player needs resources, the player will start searching and exploiting for what is needed. If the map reaches starvation (resources not available anymore, the player stops and leaves the battlefield).

Every building may also implement the Producer interface, meaning they may have their own separate threads that will produce different types of resources, once upon a build-in cooldown.

The resources they produce are established by a set of ProductionHandlers, which are implementations of a visitor interfaces hierarchy. So, based on the concrete type of Producer that tries to request production from the ProductionHandler, this one from end will call the right production method, setting the types and amount of the resources that the concrete Producer will produce (as said earlier, once upon a specific cooldown).

The updates into the player's resourceBar are thread-safe, of course, thanks to the synchronization in the player/resoucreValueWrapper classes.

Future updates coming in setting theese specific ProductionHandlers, for every producer.