This is the last commit before showing our game to the professor.
This update is bound to the last commit's promise:
"Future updates coming in setting theese specific ProductionHandlers, for every producer."

Every single building has now it's own production threat, that might be started or not, when the building is built by the player. By default, the building starts the production thread automatically. Once upon a cooldown time that is set in the Building specific Production Handler, the building will produce some specific types and amounts of resources (implemented by the game rules for example. For simplicity, every building class have theese initializations in their static blocks). The produced resources are updated by that the production threads directly to the building owner's resource bar (of course, in a syncrhonized manner, as the update/get methods in the player class are syncrhonized).

Other important updates:
-All the threads are now being safe-stopped in an asynchronous manner when clicking the "RESET" button. First, the players tell their building's production threads to stop, then the players themselves will be stopped and joined to the main Game thread.
-The global chat log might be checked for any update in the game. (exploatation/building events, resources production evenets, and also the finish event of the players that occurs when one of them is winning!